==> 1_download_data.py <==
#!/usr/bin/env python2
# -*- coding: utf-8 -*-
"""
Created on Sun Aug 12 15:18:16 2018

@author: wenzhiwang
"""

import requests

data_url = 'https://bittigermusicplayerdata.s3-us-west-2.amazonaws.com/list.html'

file_list_page = """
             Size  Lastmodified              Name   
       11526815 B  2017-06-15T04:50:23.000Z  1_1_search.log.tar.gz                              
        3463713 B  2017-06-15T04:50:23.000Z  1_2_search.log.tar.gz                              
        2172199 B  2017-06-15T04:50:23.000Z  1_3_search.log.tar.gz                              
        1964623 B  2017-06-15T04:50:24.000Z  1_4_search.log.tar.gz                              
       96295511 B  2017-06-15T04:50:24.000Z  20170301_play.log.gz                               
       65424123 B  2017-06-15T04:50:23.000Z  20170302_1_play.log.tar.gz                         
       49177957 B  2017-06-15T04:50:23.000Z  20170303_1_play.log.tar.gz                         
       45469556 B  2017-06-15T04:50:23.000Z  20170304_1_play.log.tar.gz                         
       42076341 B  2017-06-15T04:50:23.000Z  20170305_1_play.log.tar.gz                         
       35672100 B  2017-06-15T04:50:23.000Z  20170306_1_play.log.tar.gz                         
       34382105 B  2017-06-15T04:50:23.000Z  20170307_1_play.log.tar.gz                         
       32705814 B  2017-06-15T04:50:23.000Z  20170308_1_play.log.tar.gz                         
       31126117 B  2017-06-15T04:50:23.000Z  20170309_1_play.log.tar.gz                         
       57700015 B  2017-06-15T04:50:23.000Z  20170329_1_play.log.tar.gz                         
        2963334 B  2017-06-15T04:50:23.000Z  20170330_1_down.log.tar.gz                         
        2387720 B  2017-06-15T04:50:23.000Z  20170330_1_search.log.tar.gz                       
        5432075 B  2017-06-15T04:50:23.000Z  20170330_2_down.log.tar.gz                         
        3923647 B  2017-06-15T04:50:23.000Z  20170330_2_search.log.tar.gz                       
            269 B  2017-06-15T04:50:23.000Z  20170330_3.uids.tar.gz                             
       19002165 B  2017-06-15T04:50:23.000Z  20170330_3_down.log.tar.gz                         
      123905437 B  2017-06-15T04:50:23.000Z  20170330_3_play.log.tar.gz                         
       13622385 B  2017-06-15T04:50:23.000Z  20170330_3_search.log.tar.gz                       
        2137318 B  2017-06-15T04:50:23.000Z  20170331_1_down.log.tar.gz                         
       46973442 B  2017-06-15T04:50:24.000Z  20170331_1_play.log.tar.gz                         
        1848907 B  2017-06-15T04:50:23.000Z  20170331_1_search.log.tar.gz                       
        2717480 B  2017-06-15T04:50:23.000Z  20170331_2_down.log.tar.gz                         
       54658515 B  2017-06-15T04:50:24.000Z  20170331_2_play.log.tar.gz                         
        2301742 B  2017-06-15T04:50:23.000Z  20170331_2_search.log.tar.gz                       
        5222755 B  2017-06-15T04:50:23.000Z  20170331_3_down.log.tar.gz                         
       76083356 B  2017-06-15T04:50:23.000Z  20170331_3_play.log.tar.gz                         
        3996970 B  2017-06-15T04:50:23.000Z  20170331_3_search.log.tar.gz                       
        1754526 B  2017-06-15T04:50:23.000Z  20170401_1_down.log.tar.gz                         
       42948675 B  2017-06-15T04:50:23.000Z  20170401_1_play.log.tar.gz                         
        1595685 B  2017-06-15T04:50:23.000Z  20170401_1_search.log.tar.gz                       
        2020177 B  2017-06-15T04:50:24.000Z  20170401_2_down.log.tar.gz                         
       47239836 B  2017-06-15T04:50:23.000Z  20170401_2_play.log.tar.gz                         
        1766815 B  2017-06-15T04:50:24.000Z  20170401_2_search.log.tar.gz                       
        2710066 B  2017-06-15T04:50:24.000Z  20170401_3_down.log.tar.gz                         
       54112906 B  2017-06-15T04:50:23.000Z  20170401_3_play.log.tar.gz                         
        2199740 B  2017-06-15T04:50:23.000Z  20170401_3_search.log.tar.gz                       
        1746181 B  2017-06-15T04:50:23.000Z  20170402_1_down.log.tar.gz                         
       44017691 B  2017-06-15T04:50:23.000Z  20170402_1_play.log.tar.gz                         
        1582770 B  2017-06-15T04:50:23.000Z  20170402_1_search.log.tar.gz                       
        1979427 B  2017-06-15T04:50:23.000Z  20170402_2_down.log.tar.gz                         
       47582126 B  2017-06-15T04:50:25.000Z  20170402_2_play.log.tar.gz                         
        1702156 B  2017-06-15T04:50:23.000Z  20170402_2_search.log.tar.gz                       
        2312865 B  2017-06-15T04:50:24.000Z  20170402_3_down.log.tar.gz                         
       49317560 B  2017-06-15T04:50:23.000Z  20170402_3_play.log.tar.gz                         
        1958379 B  2017-06-15T04:50:23.000Z  20170402_3_search.log.tar.gz                       
        1485451 B  2017-06-15T04:50:23.000Z  20170403_1_down.log.tar.gz                         
       41647493 B  2017-06-15T04:50:23.000Z  20170403_1_play.log.tar.gz                         
        1441252 B  2017-06-15T04:50:24.000Z  20170403_1_search.log.tar.gz                       
        1632749 B  2017-06-15T04:50:23.000Z  20170403_2_down.log.tar.gz                         
       44357739 B  2017-06-15T04:50:23.000Z  20170403_2_play.log.tar.gz                         
        1539511 B  2017-06-15T04:50:24.000Z  20170403_2_search.log.tar.gz                       
        1803125 B  2017-06-15T04:50:23.000Z  20170403_3_down.log.tar.gz                         
        1647681 B  2017-06-15T04:50:23.000Z  20170403_3_search.log.tar.gz                       
        1452177 B  2017-06-15T04:50:23.000Z  20170404_1_down.log.tar.gz                         
       39757559 B  2017-06-15T04:50:24.000Z  20170404_1_play.log.tar.gz                         
        1375169 B  2017-06-15T04:50:23.000Z  20170404_1_search.log.tar.gz                       
        1658916 B  2017-06-15T04:50:24.000Z  20170404_2_down.log.tar.gz                         
       41047060 B  2017-06-15T04:50:23.000Z  20170404_2_play.log.tar.gz                         
        1453037 B  2017-06-15T04:50:23.000Z  20170404_2_search.log.tar.gz                       
        1713392 B  2017-06-15T04:50:23.000Z  20170404_3_down.log.tar.gz                         
       41266932 B  2017-06-15T04:50:24.000Z  20170404_3_play.log.tar.gz                         
        1583511 B  2017-06-15T04:50:23.000Z  20170404_3_search.log.tar.gz                       
        1091445 B  2017-06-15T04:50:24.000Z  20170405_1_down.log.tar.gz                         
       31289599 B  2017-06-15T04:50:23.000Z  20170405_1_play.log.tar.gz                         
        1151898 B  2017-06-15T04:50:24.000Z  20170405_1_search.log.tar.gz                       
        1248721 B  2017-06-15T04:50:24.000Z  20170405_2_down.log.tar.gz                         
       32271675 B  2017-06-15T04:50:24.000Z  20170405_2_play.log.tar.gz                         
        1148099 B  2017-06-15T04:50:24.000Z  20170405_2_search.log.tar.gz                       
        1278328 B  2017-06-15T04:50:23.000Z  20170405_3_down.log.tar.gz                         
       33242690 B  2017-06-15T04:50:24.000Z  20170405_3_play.log.tar.gz                         
        1235609 B  2017-06-15T04:50:23.000Z  20170405_3_search.log.tar.gz                       
         982271 B  2017-06-15T04:50:24.000Z  20170406_1_down.log.tar.gz                         
       31575031 B  2017-06-15T04:50:23.000Z  20170406_1_play.log.tar.gz                         
        1064697 B  2017-06-15T04:50:24.000Z  20170406_1_search.log.tar.gz                       
        1156266 B  2017-06-15T04:50:23.000Z  20170406_2_down.log.tar.gz                         
       32535953 B  2017-06-15T04:50:23.000Z  20170406_2_play.log.tar.gz                         
        1097103 B  2017-06-15T04:50:23.000Z  20170406_2_search.log.tar.gz                       
        1203814 B  2017-06-15T04:50:23.000Z  20170406_3_down.log.tar.gz                         
       33513836 B  2017-06-15T04:50:23.000Z  20170406_3_play.log.tar.gz                         
        1141391 B  2017-06-15T04:50:23.000Z  20170406_3_search.log.tar.gz                       
         965943 B  2017-06-15T04:50:23.000Z  20170407_1_down.log.tar.gz                         
       31133887 B  2017-06-15T04:50:23.000Z  20170407_1_play.log.tar.gz                         
        1077929 B  2017-06-15T04:50:23.000Z  20170407_1_search.log.tar.gz                       
        1104379 B  2017-06-15T04:50:24.000Z  20170407_2_down.log.tar.gz                         
       31689635 B  2017-06-15T04:50:23.000Z  20170407_2_play.log.tar.gz                         
        1055611 B  2017-06-15T04:50:24.000Z  20170407_2_search.log.tar.gz                       
        1118869 B  2017-06-15T04:50:24.000Z  20170407_3_down.log.tar.gz                         
       32733461 B  2017-06-15T04:50:23.000Z  20170407_3_play.log.tar.gz                         
        1127393 B  2017-06-15T04:50:23.000Z  20170407_3_search.log.tar.gz                       
        1128119 B  2017-06-15T04:50:24.000Z  20170408_1_down.log.tar.gz                         
       33403418 B  2017-06-15T04:50:23.000Z  20170408_1_play.log.tar.gz                         
        1204436 B  2017-06-15T04:50:23.000Z  20170408_1_search.log.tar.gz                       
        1185143 B  2017-06-15T04:50:23.000Z  20170408_2_down.log.tar.gz                         
       34072485 B  2017-06-15T04:50:23.000Z  20170408_2_play.log.tar.gz                         
        1193928 B  2017-06-15T04:50:23.000Z  20170408_2_search.log.tar.gz                       
        1201801 B  2017-06-15T04:51:31.000Z  20170408_3_down.log.tar.gz                         
       34773694 B  2017-06-15T04:51:31.000Z  20170408_3_play.log.tar.gz                         
        1237314 B  2017-06-15T04:51:31.000Z  20170408_3_search.log.tar.gz                       
        1129060 B  2017-06-15T04:51:31.000Z  20170409_1_down.log.tar.gz                         
       32582270 B  2017-06-15T04:51:31.000Z  20170409_1_play.log.tar.gz                         
        1146958 B  2017-06-15T04:51:31.000Z  20170409_1_search.log.tar.gz                       
        1136860 B  2017-06-15T04:51:31.000Z  20170409_2_down.log.tar.gz                         
       33440321 B  2017-06-15T04:51:31.000Z  20170409_2_play.log.tar.gz                         
        1164943 B  2017-06-15T04:51:31.000Z  20170409_2_search.log.tar.gz                       
        1214147 B  2017-06-15T04:51:31.000Z  20170409_3_down.log.tar.gz                         
       33371999 B  2017-06-15T04:51:31.000Z  20170409_3_play.log.tar.gz                         
        1213229 B  2017-06-15T04:51:31.000Z  20170409_3_search.log.tar.gz                       
         891074 B  2017-06-15T04:51:31.000Z  20170410_1_down.log.tar.gz                         
       28723787 B  2017-06-15T04:51:31.000Z  20170410_1_play.log.tar.gz                         
         956151 B  2017-06-15T04:51:31.000Z  20170410_1_search.log.tar.gz                       
         863248 B  2017-06-15T04:51:31.000Z  20170410_2_down.log.tar.gz                         
       28672273 B  2017-06-15T04:51:31.000Z  20170410_2_play.log.tar.gz                         
         927999 B  2017-06-15T04:51:31.000Z  20170410_2_search.log.tar.gz                       
         902297 B  2017-06-15T04:51:31.000Z  20170410_3_down.log.tar.gz                         
       29147304 B  2017-06-15T04:51:31.000Z  20170410_3_play.log.tar.gz                         
         994823 B  2017-06-15T04:51:31.000Z  20170410_3_search.log.tar.gz                       
        1044026 B  2017-06-15T04:51:31.000Z  20170411_1_down.log.tar.gz                         
       29432034 B  2017-06-15T04:51:31.000Z  20170411_1_play.log.tar.gz                         
         959708 B  2017-06-15T04:51:31.000Z  20170411_1_search.log.tar.gz                       
         988763 B  2017-06-15T04:51:31.000Z  20170411_2_down.log.tar.gz                         
       28981640 B  2017-06-15T04:51:32.000Z  20170411_2_play.log.tar.gz                         
         916462 B  2017-06-15T04:51:31.000Z  20170411_2_search.log.tar.gz                       
         987491 B  2017-06-15T04:51:32.000Z  20170412_1_down.log.tar.gz                         
       29276015 B  2017-06-15T04:51:32.000Z  20170412_1_play.log.tar.gz                         
         894312 B  2017-06-15T04:51:32.000Z  20170412_1_search.log.tar.gz                       
        1025605 B  2017-06-15T04:51:31.000Z  20170412_2_down.log.tar.gz                         
       28934309 B  2017-06-15T04:51:31.000Z  20170412_2_play.log.tar.gz                         
         868351 B  2017-06-15T04:51:31.000Z  20170412_2_search.log.tar.gz                       
         941861 B  2017-06-15T04:51:31.000Z  20170412_3_down.log.tar.gz                         
       28966132 B  2017-06-15T04:51:31.000Z  20170412_3_play.log.tar.gz                         
         902576 B  2017-06-15T04:51:31.000Z  20170412_3_search.log.tar.gz                       
         824693 B  2017-06-15T04:51:31.000Z  20170413_1_down.log.tar.gz                         
       28562498 B  2017-06-15T04:51:32.000Z  20170413_1_play.log.tar.gz                         
         893134 B  2017-06-15T04:51:31.000Z  20170413_1_search.log.tar.gz                       
         788419 B  2017-06-15T04:51:31.000Z  20170413_2_down.log.tar.gz                         
       28554387 B  2017-06-15T04:51:31.000Z  20170413_2_play.log.tar.gz                         
         887940 B  2017-06-15T04:51:32.000Z  20170413_2_search.log.tar.gz                       
         875768 B  2017-06-15T04:51:31.000Z  20170413_3_down.log.tar.gz                         
       28453794 B  2017-06-15T04:51:31.000Z  20170413_3_play.log.tar.gz                         
         935339 B  2017-06-15T04:51:31.000Z  20170413_3_search.log.tar.gz                       
         834699 B  2017-06-15T04:51:31.000Z  20170414_1_down.log.tar.gz                         
       19794279 B  2017-06-15T04:51:31.000Z  20170414_1_play.log.tar.gz                         
         857671 B  2017-06-15T04:51:32.000Z  20170414_1_search.log.tar.gz                       
         870068 B  2017-06-15T04:51:31.000Z  20170414_2_down.log.tar.gz                         
       19938829 B  2017-06-15T04:51:31.000Z  20170414_2_play.log.tar.gz                         
         858454 B  2017-06-15T04:51:31.000Z  20170414_2_search.log.tar.gz                       
         861491 B  2017-06-15T04:51:31.000Z  20170414_3_down.log.tar.gz                         
       20165880 B  2017-06-15T04:51:32.000Z  20170414_3_play.log.tar.gz                         
         901101 B  2017-06-15T04:51:31.000Z  20170414_3_search.log.tar.gz                       
         883491 B  2017-06-15T04:51:31.000Z  20170415_1_down.log.tar.gz                         
       29775809 B  2017-06-15T04:51:32.000Z  20170415_1_play.log.tar.gz                         
         995236 B  2017-06-15T04:51:31.000Z  20170415_1_search.log.tar.gz                       
         991951 B  2017-06-15T04:51:31.000Z  20170415_2_down.log.tar.gz                         
       30207894 B  2017-06-15T04:51:31.000Z  20170415_2_play.log.tar.gz                         
         992402 B  2017-06-15T04:51:31.000Z  20170415_2_search.log.tar.gz                       
         996784 B  2017-06-15T04:51:32.000Z  20170415_3_down.log.tar.gz                         
       30598122 B  2017-06-15T04:51:32.000Z  20170415_3_play.log.tar.gz                         
        1032957 B  2017-06-15T04:51:31.000Z  20170415_3_search.log.tar.gz                       
         853387 B  2017-06-15T04:51:32.000Z  20170416_1_down.log.tar.gz                         
       29810797 B  2017-06-15T04:51:31.000Z  20170416_1_play.log.tar.gz                         
         996222 B  2017-06-15T04:51:32.000Z  20170416_1_search.log.tar.gz                       
         913501 B  2017-06-15T04:51:32.000Z  20170416_2_down.log.tar.gz                         
       30070723 B  2017-06-15T04:51:32.000Z  20170416_2_play.log.tar.gz                         
        1023099 B  2017-06-15T04:51:32.000Z  20170416_2_search.log.tar.gz                       
         901357 B  2017-06-15T04:51:31.000Z  20170416_3_down.log.tar.gz                         
       29838073 B  2017-06-15T04:51:32.000Z  20170416_3_play.log.tar.gz                         
        1013745 B  2017-06-15T04:51:31.000Z  20170416_3_search.log.tar.gz                       
         745656 B  2017-06-15T04:51:32.000Z  20170417_1_down.log.tar.gz                         
       26534621 B  2017-06-15T04:51:31.000Z  20170417_1_play.log.tar.gz                         
         847713 B  2017-06-15T04:51:31.000Z  20170417_1_search.log.tar.gz                       
         810750 B  2017-06-15T04:51:31.000Z  20170417_2_down.log.tar.gz                         
       26147102 B  2017-06-15T04:51:32.000Z  20170417_2_play.log.tar.gz                         
         824043 B  2017-06-15T04:51:31.000Z  20170417_2_search.log.tar.gz                       
         777941 B  2017-06-15T04:51:31.000Z  20170417_3_down.log.tar.gz                         
       25752249 B  2017-06-15T04:51:31.000Z  20170417_3_play.log.tar.gz                         
         832787 B  2017-06-15T04:51:32.000Z  20170417_3_search.log.tar.gz                       
         727754 B  2017-06-15T04:51:32.000Z  20170418_1_down.log.tar.gz                         
       25888052 B  2017-06-15T04:51:31.000Z  20170418_1_play.log.tar.gz                         
         827652 B  2017-06-15T04:51:32.000Z  20170418_1_search.log.tar.gz                       
         765062 B  2017-06-15T04:51:31.000Z  20170418_2_down.log.tar.gz                         
       25749873 B  2017-06-15T04:51:31.000Z  20170418_2_play.log.tar.gz                         
         836100 B  2017-06-15T04:51:32.000Z  20170418_2_search.log.tar.gz                       
         712063 B  2017-06-15T04:51:31.000Z  20170418_3_down.log.tar.gz                         
       24921786 B  2017-06-15T04:51:31.000Z  20170418_3_play.log.tar.gz                         
         799392 B  2017-06-15T04:51:31.000Z  20170418_3_search.log.tar.gz                       
         599419 B  2017-06-15T04:51:32.000Z  20170419_1_down.log.tar.gz                         
       26017835 B  2017-06-15T04:51:32.000Z  20170419_1_play.log.tar.gz                         
         832010 B  2017-06-15T04:51:32.000Z  20170419_1_search.log.tar.gz                       
         556006 B  2017-06-15T04:51:32.000Z  20170419_2_down.log.tar.gz                         
       25455284 B  2017-06-15T04:51:32.000Z  20170419_2_play.log.tar.gz                         
         812772 B  2017-06-15T04:51:31.000Z  20170419_2_search.log.tar.gz                       
         547643 B  2017-06-15T04:51:31.000Z  20170419_3_down.log.tar.gz                         
       25109160 B  2017-06-15T04:51:32.000Z  20170419_3_play.log.tar.gz                         
         813572 B  2017-06-15T04:51:31.000Z  20170419_3_search.log.tar.gz                       
         756803 B  2017-06-15T04:52:22.000Z  20170420_1_down.log.tar.gz                         
       25729150 B  2017-06-15T04:52:22.000Z  20170420_1_play.log.tar.gz                         
         838770 B  2017-06-15T04:52:22.000Z  20170420_1_search.log.tar.gz                       
         698148 B  2017-06-15T04:52:22.000Z  20170420_2_down.log.tar.gz                         
       25317932 B  2017-06-15T04:52:22.000Z  20170420_2_play.log.tar.gz                         
         791995 B  2017-06-15T04:52:22.000Z  20170420_2_search.log.tar.gz                       
         715894 B  2017-06-15T04:52:22.000Z  20170420_3_down.log.tar.gz                         
       25004994 B  2017-06-15T04:52:22.000Z  20170420_3_play.log.tar.gz                         
         777847 B  2017-06-15T04:52:22.000Z  20170420_3_search.log.tar.gz                       
         759184 B  2017-06-15T04:52:22.000Z  20170421_1_down.log.tar.gz                         
       25844851 B  2017-06-15T04:52:22.000Z  20170421_1_play.log.tar.gz                         
         821816 B  2017-06-15T04:52:22.000Z  20170421_1_search.log.tar.gz                       
         725677 B  2017-06-15T04:52:22.000Z  20170421_2_down.log.tar.gz                         
       25517588 B  2017-06-15T04:52:22.000Z  20170421_2_play.log.tar.gz                         
         819987 B  2017-06-15T04:52:22.000Z  20170421_2_search.log.tar.gz                       
         701707 B  2017-06-15T04:52:22.000Z  20170421_3_down.log.tar.gz                         
       25356305 B  2017-06-15T04:52:22.000Z  20170421_3_play.log.tar.gz                         
         813603 B  2017-06-15T04:52:22.000Z  20170421_3_search.log.tar.gz                       
         735351 B  2017-06-15T04:52:22.000Z  20170422_1_down.log.tar.gz                         
       27411391 B  2017-06-15T04:52:22.000Z  20170422_1_play.log.tar.gz                         
         777314 B  2017-06-15T04:52:22.000Z  20170422_1_search.log.tar.gz                       
         764355 B  2017-06-15T04:52:22.000Z  20170422_2_down.log.tar.gz                         
       27256516 B  2017-06-15T04:52:22.000Z  20170422_2_play.log.tar.gz                         
         762065 B  2017-06-15T04:52:22.000Z  20170422_2_search.log.tar.gz                       
         724749 B  2017-06-15T04:52:22.000Z  20170422_3_down.log.tar.gz                         
       27441134 B  2017-06-15T04:52:22.000Z  20170422_3_play.log (1).tar.gz                     
       27441134 B  2017-06-15T04:52:22.000Z  20170422_3_play.log.tar.gz                         
         793283 B  2017-06-15T04:52:22.000Z  20170422_3_search.log.tar.gz                       
         714044 B  2017-06-15T04:52:22.000Z  20170423_1_down.log.tar.gz                         
       27523910 B  2017-06-15T04:52:22.000Z  20170423_1_play.log.tar.gz                         
         903118 B  2017-06-15T04:52:22.000Z  20170423_1_search.log.tar.gz                       
         709108 B  2017-06-15T04:52:22.000Z  20170423_2_down.log.tar.gz                         
       27193776 B  2017-06-15T04:52:22.000Z  20170423_2_play.log.tar.gz                         
         884938 B  2017-06-15T04:52:22.000Z  20170423_2_search.log.tar.gz                       
         750992 B  2017-06-15T04:52:22.000Z  20170423_3_down.log.tar.gz                         
       26863772 B  2017-06-15T04:52:22.000Z  20170423_3_play.log.tar.gz                         
         919539 B  2017-06-15T04:52:22.000Z  20170423_3_search.log.tar.gz                       
         547769 B  2017-06-15T04:52:22.000Z  20170424_1_down.log.tar.gz                         
       13631488 B  2017-06-15T04:52:22.000Z  20170424_1_play.log.tar.gz                         
         777441 B  2017-06-15T04:52:22.000Z  20170424_1_search.log.tar.gz                       
         612017 B  2017-06-15T04:52:22.000Z  20170424_2_down.log.tar.gz                         
        1048576 B  2017-06-15T04:52:23.000Z  20170424_2_play.log.tar.gz                         
         772575 B  2017-06-15T04:52:22.000Z  20170424_2_search.log.tar.gz                       
         536018 B  2017-06-15T04:52:22.000Z  20170424_3_down.log.tar.gz                         
       23340101 B  2017-06-15T04:52:22.000Z  20170424_3_play.log.tar.gz                         
         759755 B  2017-06-15T04:52:22.000Z  20170424_3_search.log.tar.gz                       
         640246 B  2017-06-15T04:52:22.000Z  20170425_1_down.log.tar.gz                         
       24531214 B  2017-06-15T04:52:22.000Z  20170425_1_play.log.tar.gz                         
         576490 B  2017-06-15T04:52:22.000Z  20170425_1_search.log.tar.gz                       
         620400 B  2017-06-15T04:52:22.000Z  20170425_2_down.log.tar.gz                         
       23155158 B  2017-06-15T04:52:22.000Z  20170425_2_play.log.tar.gz                         
         564900 B  2017-06-15T04:52:22.000Z  20170425_2_search.log.tar.gz                       
         593137 B  2017-06-15T04:52:22.000Z  20170425_3_down.log.tar.gz                         
       23503687 B  2017-06-15T04:52:22.000Z  20170425_3_play.log.tar.gz                         
         571926 B  2017-06-15T04:52:22.000Z  20170425_3_search.log.tar.gz                       
         591608 B  2017-06-15T04:52:22.000Z  20170426_1_down.log.tar.gz                         
       23988275 B  2017-06-15T04:52:22.000Z  20170426_1_play.log.tar.gz                         
         420470 B  2017-06-15T04:52:23.000Z  20170426_1_search.log.tar.gz                       
         629058 B  2017-06-15T04:52:22.000Z  20170426_2_down.log.tar.gz                         
       23355484 B  2017-06-15T04:52:22.000Z  20170426_2_play.log.tar.gz                         
         415918 B  2017-06-15T04:52:22.000Z  20170426_2_search.log.tar.gz                       
         571250 B  2017-06-15T04:52:22.000Z  20170426_3_down.log.tar.gz                         
       23093713 B  2017-06-15T04:52:22.000Z  20170426_3_play.log.tar.gz                         
         407692 B  2017-06-15T04:52:22.000Z  20170426_3_search.log.tar.gz                       
         657363 B  2017-06-15T04:52:22.000Z  20170427_1_down.log.tar.gz                         
       24061905 B  2017-06-15T04:52:22.000Z  20170427_1_play.log.tar.gz                         
         400976 B  2017-06-15T04:52:22.000Z  20170427_1_search.log.tar.gz                       
         648391 B  2017-06-15T04:52:22.000Z  20170427_2_down.log.tar.gz                         
       23481868 B  2017-06-15T04:52:22.000Z  20170427_2_play.log.tar.gz                         
         397801 B  2017-06-15T04:52:22.000Z  20170427_2_search.log.tar.gz                       
         641470 B  2017-06-15T04:52:22.000Z  20170427_3_down.log.tar.gz                         
       31323865 B  2017-06-15T04:52:22.000Z  20170427_3_play.log.tar.gz                         
         393410 B  2017-06-15T04:52:22.000Z  20170427_3_search.log.tar.gz                       
         703393 B  2017-06-15T04:52:22.000Z  20170428_1_down.log.tar.gz                         
       24592387 B  2017-06-15T04:52:22.000Z  20170428_1_play.log.tar.gz                         
         533969 B  2017-06-15T04:52:22.000Z  20170428_1_search.log.tar.gz                       
         673533 B  2017-06-15T04:52:22.000Z  20170428_2_down.log.tar.gz                         
       24539641 B  2017-06-15T04:52:22.000Z  20170428_2_play.log.tar.gz                         
         556982 B  2017-06-15T04:52:22.000Z  20170428_2_search.log.tar.gz                       
         665884 B  2017-06-15T04:52:22.000Z  20170428_3_down.log.tar.gz                         
       24336288 B  2017-06-15T04:52:22.000Z  20170428_3_play.log.tar.gz                         
         541331 B  2017-06-15T04:52:22.000Z  20170428_3_search.log.tar.gz                       
         802164 B  2017-06-15T04:52:22.000Z  20170429_1_down.log.tar.gz                         
       26564661 B  2017-06-15T04:52:22.000Z  20170429_1_play.log.tar.gz                         
         851038 B  2017-06-15T04:52:22.000Z  20170429_1_search.log.tar.gz                       
         775926 B  2017-06-15T04:52:22.000Z  20170429_2_down.log.tar.gz                         
       27343684 B  2017-06-15T04:52:22.000Z  20170429_2_play.log.tar.gz                         
         853661 B  2017-06-15T04:52:22.000Z  20170429_2_search.log.tar.gz                       
         825371 B  2017-06-15T04:52:22.000Z  20170429_3_down.log.tar.gz                         
       27297892 B  2017-06-15T04:52:22.000Z  20170429_3_play.log.tar.gz                         
         859056 B  2017-06-15T04:52:22.000Z  20170429_3_search.log.tar.gz                       
         708962 B  2017-06-15T04:52:22.000Z  20170430_1_down.log.tar.gz                         
       25585694 B  2017-06-15T04:52:22.000Z  20170430_1_play.log.tar.gz                         
         857577 B  2017-06-15T04:52:22.000Z  20170430_1_search.log.tar.gz                       
         750111 B  2017-06-15T04:52:22.000Z  20170430_2_down.log.tar.gz                         
       26182661 B  2017-06-15T04:52:22.000Z  20170430_2_play.log.tar.gz                         
         874173 B  2017-06-15T04:52:22.000Z  20170430_2_search.log.tar.gz                       
         753777 B  2017-06-15T04:52:22.000Z  20170430_3_down.log.tar.gz                         
       26262727 B  2017-06-15T04:52:22.000Z  20170430_3_play.log.tar.gz                         
         855085 B  2017-06-15T04:53:08.000Z  20170430_3_search.log.tar.gz                       
         894441 B  2017-06-15T04:53:08.000Z  20170501_1_down.log.tar.gz                         
       24931945 B  2017-06-15T04:53:08.000Z  20170501_1_play.log.tar.gz                         
         820621 B  2017-06-15T04:53:08.000Z  20170501_1_search.log.tar.gz                       
         885225 B  2017-06-15T04:53:08.000Z  20170501_2_down.log.tar.gz                         
       24499350 B  2017-06-15T04:53:08.000Z  20170501_2_play.log.tar.gz                         
         815892 B  2017-06-15T04:53:08.000Z  20170501_2_search.log.tar.gz                       
         739808 B  2017-06-15T04:53:08.000Z  20170501_3_down.log.tar.gz                         
       24807163 B  2017-06-15T04:53:08.000Z  20170501_3_play.log.tar.gz                         
         810381 B  2017-06-15T04:53:08.000Z  20170501_3_search.log.tar.gz                       
         561394 B  2017-06-15T04:53:08.000Z  20170502_1_down.log.tar.gz                         
       21568540 B  2017-06-15T04:53:08.000Z  20170502_1_play.log.tar.gz                         
         702938 B  2017-06-15T04:53:08.000Z  20170502_1_search.log.tar.gz                       
         561836 B  2017-06-15T04:53:08.000Z  20170502_2_down.log.tar.gz                         
       20844120 B  2017-06-15T04:53:08.000Z  20170502_2_play.log.tar.gz                         
         662112 B  2017-06-15T04:53:08.000Z  20170502_2_search.log.tar.gz                       
         571301 B  2017-06-15T04:53:08.000Z  20170502_3_down.log.tar.gz                         
       20574309 B  2017-06-15T04:53:08.000Z  20170502_3_play.log.tar.gz                         
         659819 B  2017-06-15T04:53:08.000Z  20170502_3_search.log.tar.gz                       
         536086 B  2017-06-15T04:53:08.000Z  20170503_1_down.log.tar.gz                         
       22103656 B  2017-06-15T04:53:08.000Z  20170503_1_play.log.tar.gz                         
         694544 B  2017-06-15T04:53:08.000Z  20170503_1_search.log.tar.gz                       
         501305 B  2017-06-15T04:53:08.000Z  20170503_2_down.log.tar.gz                         
       21381047 B  2017-06-15T04:53:08.000Z  20170503_2_play.log.tar.gz                         
         676893 B  2017-06-15T04:53:08.000Z  20170503_2_search.log.tar.gz                       
         523371 B  2017-06-15T04:53:08.000Z  20170503_3_down.log.tar.gz                         
       21098875 B  2017-06-15T04:53:08.000Z  20170503_3_play.log.tar.gz                         
         660165 B  2017-06-15T04:53:08.000Z  20170503_3_search.log.tar.gz                       
         615638 B  2017-06-15T04:53:08.000Z  20170504_1_down.log.tar.gz                         
       21690250 B  2017-06-15T04:53:08.000Z  20170504_1_play.log.tar.gz                         
         689176 B  2017-06-15T04:53:08.000Z  20170504_1_search.log.tar.gz                       
         608204 B  2017-06-15T04:53:08.000Z  20170504_2_down.log.tar.gz                         
       21125373 B  2017-06-15T04:53:08.000Z  20170504_2_play.log.tar.gz                         
         664625 B  2017-06-15T04:53:08.000Z  20170504_2_search.log.tar.gz                       
         513607 B  2017-06-15T04:53:08.000Z  20170504_3_down.log.tar.gz                         
       20949397 B  2017-06-15T04:53:08.000Z  20170504_3_play.log.tar.gz                         
         657321 B  2017-06-15T04:53:08.000Z  20170504_3_search.log.tar.gz                       
         673423 B  2017-06-15T04:53:08.000Z  20170505_1_down.log.tar.gz                         
       22359294 B  2017-06-15T04:53:08.000Z  20170505_1_play.log.tar.gz                         
         722158 B  2017-06-15T04:53:08.000Z  20170505_1_search.log.tar.gz                       
         587864 B  2017-06-15T04:53:09.000Z  20170505_2_down.log.tar.gz                         
       21750093 B  2017-06-15T04:53:08.000Z  20170505_2_play.log.tar.gz                         
         688561 B  2017-06-15T04:53:08.000Z  20170505_2_search.log.tar.gz                       
         543156 B  2017-06-15T04:53:08.000Z  20170505_3_down.log.tar.gz                         
       21232254 B  2017-06-15T04:53:08.000Z  20170505_3_play.log.tar.gz                         
         663460 B  2017-06-15T04:53:08.000Z  20170505_3_search.log.tar.gz                       
         735911 B  2017-06-15T04:53:08.000Z  20170506_1_down.log.tar.gz                         
       24563320 B  2017-06-15T04:53:08.000Z  20170506_1_play.log.tar.gz                         
         793474 B  2017-06-15T04:53:08.000Z  20170506_1_search.log.tar.gz                       
         728237 B  2017-06-15T04:53:08.000Z  20170506_2_down.log.tar.gz                         
       24472302 B  2017-06-15T04:53:08.000Z  20170506_2_play.log.tar.gz                         
         791825 B  2017-06-15T04:53:08.000Z  20170506_2_search.log.tar.gz                       
         713156 B  2017-06-15T04:53:08.000Z  20170506_3_down.log.tar.gz                         
       23745209 B  2017-06-15T04:53:08.000Z  20170506_3_play.log.tar.gz                         
         785464 B  2017-06-15T04:53:08.000Z  20170506_3_search.log.tar.gz                       
         622904 B  2017-06-15T04:53:08.000Z  20170507_1_down.log.tar.gz                         
       23991815 B  2017-06-15T04:53:08.000Z  20170507_1_play.log.tar.gz                         
         797841 B  2017-06-15T04:53:08.000Z  20170507_1_search.log.tar.gz                       
         675298 B  2017-06-15T04:53:08.000Z  20170507_2_down.log.tar.gz                         
       23757712 B  2017-06-15T04:53:08.000Z  20170507_2_play.log.tar.gz                         
         781077 B  2017-06-15T04:53:08.000Z  20170507_2_search.log.tar.gz                       
         641197 B  2017-06-15T04:53:08.000Z  20170507_3_down.log.tar.gz                         
       23056350 B  2017-06-15T04:53:08.000Z  20170507_3_play.log.tar.gz                         
         759802 B  2017-06-15T04:53:08.000Z  20170507_3_search.log.tar.gz                       
         543029 B  2017-06-15T04:53:08.000Z  20170508_1_down.log.tar.gz                         
       21753794 B  2017-06-15T04:53:08.000Z  20170508_1_play.log.tar.gz                         
         662705 B  2017-06-15T04:53:08.000Z  20170508_1_search.log.tar.gz                       
         567977 B  2017-06-15T04:53:12.000Z  20170508_2_down.log.tar.gz                         
       21386777 B  2017-06-15T04:53:08.000Z  20170508_2_play.log.tar.gz                         
         674471 B  2017-06-15T04:53:08.000Z  20170508_2_search.log.tar.gz                       
         590985 B  2017-06-15T04:53:08.000Z  20170508_3_down.log.tar.gz                         
       20962683 B  2017-06-15T04:53:08.000Z  20170508_3_play.log.tar.gz                         
         646467 B  2017-06-15T04:53:08.000Z  20170508_3_search.log.tar.gz                       
         543030 B  2017-06-15T04:53:08.000Z  20170509_1_down.log.tar.gz                         
       21690550 B  2017-06-15T04:53:08.000Z  20170509_1_play.log.tar.gz                         
         662706 B  2017-06-15T04:53:08.000Z  20170509_1_search.log.tar.gz                       
         567981 B  2017-06-15T04:53:08.000Z  20170509_2_down.log.tar.gz                         
       21220407 B  2017-06-15T04:53:08.000Z  20170509_2_play.log.tar.gz                         
         674471 B  2017-06-15T04:53:08.000Z  20170509_2_search.log.tar.gz                       
         590989 B  2017-06-15T04:53:08.000Z  20170509_3_down.log.tar.gz                         
       20370631 B  2017-06-15T04:53:08.000Z  20170509_3_play.log.tar.gz                         
         646467 B  2017-06-15T04:53:08.000Z  20170509_3_search.log.tar.gz                       
         537830 B  2017-06-15T04:53:08.000Z  20170510_1_down.log.tar.gz                         
       21466828 B  2017-06-15T04:53:08.000Z  20170510_1_play.log.tar.gz                         
         649408 B  2017-06-15T04:53:08.000Z  20170510_1_search.log.tar.gz                       
         552921 B  2017-06-15T04:53:08.000Z  20170510_2_down.log.tar.gz                         
       21213582 B  2017-06-15T04:53:08.000Z  20170510_2_play.log.tar.gz                         
         642448 B  2017-06-15T04:53:08.000Z  20170510_2_search.log.tar.gz                       
         538143 B  2017-06-15T04:53:08.000Z  20170510_3_down.log.tar.gz                         
       20461273 B  2017-06-15T04:53:08.000Z  20170510_3_play.log.tar.gz                         
         642702 B  2017-06-15T04:53:08.000Z  20170510_3_search.log.tar.gz                       
         481476 B  2017-06-15T04:53:08.000Z  20170511_1_down.log.tar.gz                         
       21223376 B  2017-06-15T04:53:08.000Z  20170511_1_play.log.tar.gz                         
         679512 B  2017-06-15T04:53:08.000Z  20170511_1_search.log.tar.gz                       
         482949 B  2017-06-15T04:53:08.000Z  20170511_2_down.log.tar.gz                         
       20594616 B  2017-06-15T04:53:08.000Z  20170511_2_play.log.tar.gz                         
         622609 B  2017-06-15T04:53:08.000Z  20170511_2_search.log.tar.gz                       
         492245 B  2017-06-15T04:53:08.000Z  20170511_3_down.log.tar.gz                         
       20347821 B  2017-06-15T04:53:08.000Z  20170511_3_play.log.tar.gz                         
         623879 B  2017-06-15T04:53:55.000Z  20170511_3_search.log.tar.gz                       
         499184 B  2017-06-15T04:53:55.000Z  20170512_1_down.log.tar.gz                         
       21691240 B  2017-06-15T04:53:55.000Z  20170512_1_play.log.tar.gz                         
         697372 B  2017-06-15T04:53:55.000Z  20170512_1_search.log.tar.gz                       
         554060 B  2017-06-15T04:53:55.000Z  20170512_2_down.log.tar.gz                         
       21557009 B  2017-06-15T04:53:55.000Z  20170512_2_play.log.tar.gz                         
         686030 B  2017-06-15T04:53:55.000Z  20170512_2_search.log.tar.gz                       
         544918 B  2017-06-15T04:53:55.000Z  20170512_3_down.log.tar.gz                         
       20873703 B  2017-06-15T04:53:55.000Z  20170512_3_play.log.tar.gz                         
         666689 B  2017-06-15T04:53:55.000Z  20170512_3_search.log.tar.gz                       
        3258024 B  2017-06-15T04:53:55.000Z  3_1.uids.gz                                        
           6032 B  2017-06-15T04:44:51.000Z  list.html                                          

 Amazon S3 Bucket list v 1.1
"""


file_list = [row.strip().split()[-1] for row in file_list_page.split('\n') if ('search' in row or 'play' in row or 'down' in row)]


file_list = [row for row in file_list if row[:8] >= '20170330']




def download_file(url, local_filename):
    
    # note the stream = true parameter
    
    r = requests.get(url, stream = True)
    
    with open(local_filename, 'wb') as f:
        
        for chunk in r.iter_content(chunk_size = 1024) :
            
            if chunk:# filter out keep-alive new chunks
                
                f.write(chunk)
                #f.flush() commented by recommendation from J.F.Sebastian
    return local_filename


base_url = 'https://bittigermusicplayerdata.s3-us-west-2.amazonaws.com/'

target_dir = '../data/raw/'


for file in file_list:
    
    download_file(base_url+file, target_dir + file)
